# backend/docs/swagger.yaml
definitions:
  model.Card:
    properties:
      created_by:
        type: string
      created_on:
        type: integer
      deleted_on:
        type: integer
      description:
        type: string
      effect:
        type: string
      game_id:
        type: integer
      id:
        type: integer
      is_del:
        type: integer
      modified_by:
        type: string
      modified_on:
        type: integer
      name:
        type: string
      type:
        description: 'Added: role, event, item'
        type: string
    type: object
  model.Game:
    properties:
      card_count:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      created_on:
        type: integer
      deleted_on:
        type: integer
      description:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      modified_by:
        type: string
      modified_on:
        type: integer
      style:
        type: string
      theme:
        type: string
    type: object
  v1.GameResponse:
    properties:
      card_count:
        type: integer
      cards:
        items:
          $ref: '#/definitions/model.Card'
        type: array
      description:
        type: string
      id:
        type: integer
      style:
        type: string
      theme:
        type: string
    type: object
  v1.GenerateGameRequest:
    properties:
      cardCount:
        maximum: 100
        minimum: 10
        type: integer
      description:
        type: string
      style:
        type: string
      theme:
        type: string
    required:
    - cardCount
    - style
    - theme
    type: object
info:
  contact: {}
paths:
  /api/v1/game:
    post:
      consumes:
      - application/json
      description: Generates a new board game using Gemini AI based on the provided
        theme, card count, style, and optional description. The endpoint creates a
        game record, generates cards, and stores related metadata.
      parameters:
      - description: Game generation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.GenerateGameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Game generated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Quota exceeded
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Generate a new board game
      tags:
      - game
  /api/v1/game/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a game and its cards by the provided game ID.
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GameResponse'
        "404":
          description: game not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: failed to fetch cards
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get game by ID
      tags:
      - game
  /api/v1/games:
    get:
      description: Retrieves all games that are not marked as deleted.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Game'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List games
      tags:
      - games
  /api/v1/games/{id}/pdf:
    get:
      description: Generates a PDF file containing the board game's details and its
        cards, and returns the PDF file.
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: PDF file
          schema:
            type: file
        "404":
          description: game not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Generate PDF for a board game
      tags:
      - games
swagger: "2.0"